move to real a src file, and minor code quality refactor
